cambio en test game-handler y valid_move ahora en game
